refactor(analysis)!: move pattern detector from sonda CLI to sonda-core::analysis
`pattern.rs` is pure statistical analysis — takes `Vec<f64>`, returns a
classified `Pattern`, has zero CSV-reader or CLI dependencies. It belongs
in the engine, not in the CLI crate.

This move sets up 1.9c's `sonda new --from <csv>` to call into the
pattern detector via `sonda_core::analysis::pattern::detect_pattern`
without depending on the soon-to-be-deleted `sonda/src/import/` tree.

  sonda/src/import/pattern.rs  →  sonda-core/src/analysis/pattern.rs

Two callers rewired:

  - `sonda/src/import/mod.rs`     `import::pattern::` → `sonda_core::analysis::pattern::`
  - `sonda/src/import/yaml_gen.rs` `super::pattern::` → `sonda_core::analysis::pattern::`
  - `sonda/src/init/mod.rs`       `import::pattern::` → `sonda_core::analysis::pattern::`
                                   (14 references — all mechanical)

`sonda-core/src/scenarios/mod.rs` deletion stays deferred to 1.9c —
all its consumers (`catalog.rs`, `scenarios.rs`, etc.) are CLI catalog
code being deleted alongside it.

Second of five sub-PRs composing the 1.9 release. Stacks on 1.9a.
PR base is release/1.9.

BREAKING CHANGE: `sonda_core::analysis::pattern::Pattern` is now a new
public type. The old `sonda::import::pattern::Pattern` path is gone.
Internal use only — no external crates depended on the CLI-crate path.
