+++ sonda-core/src/analysis/mod.rs
@@ -1,0 +1,3 @@
+//! Statistical analysis utilities — pattern detection from time-series data.
+
+pub mod pattern;
