+++ sonda-core/src/analysis/pattern.rs
